fix restricted debug profile
restricted debug profile set runAsNonRoot but didn't set a runAsUser,
and this upsets the kubelet, causing the debug container to not be able
to start with a "container has runAsNonRoot and image will run as root"
error.

While I was at it, I also noticed that restricted debug profile clears
security context in the node debug case, which is a problem, because
by clearing the security context, we would have lost the run-as-root
and drop-all-capabilities properties needed for the restricted profile.